create gpu-registry
