+++ src/render/mesh-pool.ts
@@ -46,6 +46,9 @@ export interface MeshHandle<U extends CyStructDesc> {
 
   readonly readonlyMesh?: Mesh;
 
+  // state
+  enabled: boolean;
+
   // used as the uniform for this mesh
   shaderData: CyToTS<U>;
 }
@@ -184,6 +187,7 @@ export function createMeshPool<V extends CyStructDesc, U extends CyStructDesc>(
 
     const handle: MeshHandle<U> = {
       mId: nextMeshId++,
+      enabled: true,
       triNum: m.tri.length,
       lineNum: m.lines?.length ?? 0,
       vertNum: m.pos.length,
