+++ src/render/render_init.ts
@@ -3,7 +3,7 @@ import { EM, EntityManager } from "../entity-manager.js";
 import { FORCE_WEBGL } from "../main.js";
 import { MAX_MESHES, MAX_VERTICES } from "./pipelines.js";
 import { RenderableConstructDef, Renderer } from "./renderer.js";
-import { attachToCanvasWebgl } from "./render_webgl.js";
+// import { attachToCanvasWebgl } from "./render_webgl.js";
 import { createWebGPURenderer } from "./render_webgpu.js";
 
 // TODO(@darzu): ECS this
@@ -53,8 +53,9 @@ async function init(
       }
     }
   }
-  if (!rendererInit)
-    rendererInit = attachToCanvasWebgl(canvas, MAX_MESHES, MAX_VERTICES);
+  // TODO(@darzu): re-enable WebGL
+  // if (!rendererInit)
+  //   rendererInit = attachToCanvasWebgl(canvas, MAX_MESHES, MAX_VERTICES);
   if (!rendererInit) throw "Unable to create webgl or webgpu renderer";
   console.log(`Renderer: ${usingWebGPU ? "webGPU" : "webGL"}`);
